+++ apps/frontend/src/app/3_organisms/index.tsx
@@ -1,3 +1,4 @@
 export * from './Header/Header';
 export * from './TransactionStepDialog';
+export * from './TransactionHistoryFrame';
 export * from './Footer/Footer';