+++ apps/frontend/src/utils/constants.ts
@@ -7,8 +7,8 @@ export const sovrynLinks = {
   telegram: 'https://t.me/SovrynBitcoin',
   twitter: 'https://twitter.com/Sovrynbtc',
   github: 'https://github.com/DistributedCollective',
-  wiki: 'https://wiki.sovryn.app/en/',
-  dappAlpha: 'https://live.sovryn.app/',
+  wiki: 'https://wiki.sovryn.app/en',
+  dappAlpha: 'https://live.sovryn.app',
   blog: 'https://www.sovryn.app/blog',
 };
 
@@ -32,6 +32,8 @@ export const CR_THRESHOLDS = {
 
 export const graphRskUrl = process.env.REACT_APP_GRAPH_RSK;
 
+export const graphZeroUrl = process.env.REACT_APP_GRAPH_ZERO;
+
 export const servicesConfig = {
   [Environments.Testnet]: {
     [Services.Maintenance]:
@@ -43,3 +45,5 @@ export const servicesConfig = {
     [Services.Notification]: 'https://notify.sovryn.app/',
   },
 };
+
+export const EXPORT_RECORD_LIMIT = 500;